+++ tests/testthat/test-op-square.R
@@ -1,0 +1,5 @@
+test_that("basic tests", {
+  hlo_test_uni(hlo_square, function(x) x * x)
+})
+
+# Errors are tested in test-type_inference.R (via infer_types_float_uni)